requirements.txt
django==5.1.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
google-auth==2.48.0
anthropic==0.79.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.3
